Converting intersectsOtherOperation to a free template function intersects.
